[TECH-556] Move versioning logic to projects namespace
because we're talking about project (schema) versions

branch: feature/TECH-556-config-upgrades
